+++ CHANGELOG.md
@@ -3,6 +3,10 @@
 # IntelliJ CUE Language Support Changelog
 
 ## [Unreleased]
+- Setting for highlighting
+- Highlighting setting for optional field names
+
+## [0.3.1]
 - Improved support for the CUE grammar
 - Improved highlighting